* Lots of compile fixes
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@3725 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ configmanager.cpp
@@ -18,7 +18,7 @@
 #include <iostream>
 
 #include "configmanager.h"
-#include "houses.h"
+#include "house.h"
 #include "tools.h"
 
 ConfigManager::ConfigManager()
@@ -323,7 +323,7 @@ bool ConfigManager::reload()
 	if(!load())
 		return false;
 
-	if(m_confNumber[HOUSE_PRICE] == tmp)
+	if((uint32_t)m_confNumber[HOUSE_PRICE] == tmp)
 		return true;
 
 	for(HouseMap::iterator it = Houses::getInstance()->getHouseBegin();
